requirements.txt
streamlit
python-dotenv
pypdf
qdrant-client
sentence-transformers
transformers
langchain
langchain-community
langchain-ollama
